refactor: 이벤트 조회 API 어드민용, 참가자용 분리
